MB-4226 Rename internal folders to payloads and move files up one level
